Confirmation service

+++ modules/core/client/controllers/sensors/sensors.client.controller.js
@@ -1,6 +1,7 @@
 'use strict';
 
-angular.module('core').controller('SensorsController', ['$scope', '$http', '$modal', '$log', function ($scope, $http, $modal, $log) {
+angular.module('core')
+    .controller('SensorsController', ['$scope', '$http', '$modal', '$log', 'confirmation', function ($scope, $http, $modal, $log, confirmation) {
     $scope.sensors = [];
 
     $scope.add = function(){
@@ -35,12 +36,14 @@ angular.module('core').controller('SensorsController', ['$scope', '$http', '$mod
     };
 
     $scope.delete = function(sensor, index) {
-        $http({
-            method: 'DELETE',
-            url: '/api/sensors/' + sensor._id
-        }).then(function() {
-            $scope.sensors.splice(index, 1);
-        }, errorCallback);
+        confirmation.confirm('Warning!', 'Do you really want to delete this item?', function() {
+            $http({
+                method: 'DELETE',
+                url: '/api/sensors/' + sensor._id
+            }).then(function() {
+                $scope.sensors.splice(index, 1);
+            }, errorCallback);
+        });
     };
 
     $http({
